+++ samplers/samplers.go
@@ -23,6 +23,23 @@ const (
 	GaugeMetric
 )
 
+// RouteInformation is a key-only map indicating sink names that are
+// supposed to receive a metric. A nil RouteInformation value
+// corresponds to the "every sink" value; an entry in a non-nil
+// RouteInformation means that the key should receive the metric.
+type RouteInformation map[string]struct{}
+
+// RouteTo returns true if the named sink should receive a metric
+// according to the route table. A nil route table causes any sink to
+// be eligible for the metric.
+func (ri RouteInformation) RouteTo(name string) bool {
+	if ri == nil {
+		return true
+	}
+	_, ok := ri[name]
+	return ok
+}
+
 // InterMetric represents a metric that has been completed and is ready for
 // flushing by sinks.
 type InterMetric struct {
@@ -31,6 +48,11 @@ type InterMetric struct {
 	Value     float64
 	Tags      []string
 	Type      MetricType
+
+	// Sinks, if non-nil, indicates which metric sinks a metric
+	// should be inserted into. If nil, that means the metric is
+	// meant to go to every sink.
+	Sinks RouteInformation
 }
 
 type Aggregate int
@@ -80,6 +102,25 @@ type JSONMetric struct {
 	Value []byte `json:"value"`
 }
 
+const sinkPrefix string = "veneursinkonly:"
+
+func routeInfo(tags []string) RouteInformation {
+	var info RouteInformation
+	for _, tag := range tags {
+		if !strings.HasPrefix(tag, sinkPrefix) {
+			continue
+		}
+		if info == nil {
+			info = make(RouteInformation)
+		}
+		// Take the tag suffix (the part after the ':' in
+		// "veneursinkonly:", and make that the key in our
+		// route information map:
+		info[tag[len(sinkPrefix):]] = struct{}{}
+	}
+	return info
+}
+
 // Counter is an accumulator
 type Counter struct {
 	Name  string
@@ -102,6 +143,7 @@ func (c *Counter) Flush(interval time.Duration) []InterMetric {
 		Value:     float64(c.value),
 		Tags:      tags,
 		Type:      CounterMetric,
+		Sinks:     routeInfo(tags),
 	}}
 }
 
@@ -167,6 +209,7 @@ func (g *Gauge) Flush() []InterMetric {
 		Value:     float64(g.value),
 		Tags:      tags,
 		Type:      GaugeMetric,
+		Sinks:     routeInfo(tags),
 	}}
 }
 
@@ -245,6 +288,7 @@ func (s *Set) Flush() []InterMetric {
 		Value:     float64(s.Hll.Estimate()),
 		Tags:      tags,
 		Type:      GaugeMetric,
+		Sinks:     routeInfo(tags),
 	}}
 }
 
@@ -329,6 +373,7 @@ func NewHist(Name string, Tags []string) *Histo {
 func (h *Histo) Flush(interval time.Duration, percentiles []float64, aggregates HistogramAggregates) []InterMetric {
 	now := time.Now().Unix()
 	metrics := make([]InterMetric, 0, aggregates.Count+len(percentiles))
+	sinks := routeInfo(h.Tags)
 
 	if (aggregates.Value&AggregateMax) == AggregateMax && !math.IsInf(h.LocalMax, 0) {
 		// Defensively recopy tags to avoid aliasing bugs in case multiple InterMetrics share the same
@@ -341,6 +386,7 @@ func (h *Histo) Flush(interval time.Duration, percentiles []float64, aggregates 
 			Value:     float64(h.LocalMax),
 			Tags:      tags,
 			Type:      GaugeMetric,
+			Sinks:     sinks,
 		})
 	}
 	if (aggregates.Value&AggregateMin) == AggregateMin && !math.IsInf(h.LocalMin, 0) {
@@ -352,6 +398,7 @@ func (h *Histo) Flush(interval time.Duration, percentiles []float64, aggregates 
 			Value:     float64(h.LocalMin),
 			Tags:      tags,
 			Type:      GaugeMetric,
+			Sinks:     sinks,
 		})
 	}
 
@@ -364,6 +411,7 @@ func (h *Histo) Flush(interval time.Duration, percentiles []float64, aggregates 
 			Value:     float64(h.LocalSum),
 			Tags:      tags,
 			Type:      GaugeMetric,
+			Sinks:     sinks,
 		})
 	}
 
@@ -378,6 +426,7 @@ func (h *Histo) Flush(interval time.Duration, percentiles []float64, aggregates 
 			Value:     float64(h.LocalSum / h.LocalWeight),
 			Tags:      tags,
 			Type:      GaugeMetric,
+			Sinks:     sinks,
 		})
 	}
 
@@ -393,6 +442,7 @@ func (h *Histo) Flush(interval time.Duration, percentiles []float64, aggregates 
 			Value:     float64(h.LocalWeight),
 			Tags:      tags,
 			Type:      CounterMetric,
+			Sinks:     sinks,
 		})
 	}
 
@@ -407,6 +457,7 @@ func (h *Histo) Flush(interval time.Duration, percentiles []float64, aggregates 
 				Value:     float64(h.Value.Quantile(0.5)),
 				Tags:      tags,
 				Type:      GaugeMetric,
+				Sinks:     sinks,
 			},
 		)
 	}
@@ -422,6 +473,7 @@ func (h *Histo) Flush(interval time.Duration, percentiles []float64, aggregates 
 			Value:     float64(h.LocalWeight / h.LocalReciprocalSum),
 			Tags:      tags,
 			Type:      GaugeMetric,
+			Sinks:     sinks,
 		})
 	}
 
@@ -437,6 +489,7 @@ func (h *Histo) Flush(interval time.Duration, percentiles []float64, aggregates 
 				Value:     float64(h.Value.Quantile(p)),
 				Tags:      tags,
 				Type:      GaugeMetric,
+				Sinks:     sinks,
 			},
 		)
 	}